pyproject.toml
[tool.poetry]
name = "deeqnlpy"
version = "0.9.0"
description = "The deeq nlp python client library"
authors = ["Gihyun YUN <gih2yun@baikal.ai>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.6"
grpcio = "1.35.0"
protobuf = "3.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
